Decomposition of code in FragmentDialect
