+++ rt/ws/transfer/src/main/java/org/apache/cxf/ws/transfer/dialect/fragment/FragmentDialect.java
@@ -19,15 +19,9 @@
 
 package org.apache.cxf.ws.transfer.dialect.fragment;
 
-import java.util.Iterator;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.util.HashMap;
+import java.util.Map;
 import javax.xml.bind.JAXBElement;
-import javax.xml.namespace.NamespaceContext;
-import javax.xml.xpath.XPath;
-import javax.xml.xpath.XPathConstants;
-import javax.xml.xpath.XPathException;
-import javax.xml.xpath.XPathExpressionException;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
@@ -37,8 +31,10 @@ import org.apache.cxf.ws.transfer.Get;
 import org.apache.cxf.ws.transfer.Put;
 import org.apache.cxf.ws.transfer.Representation;
 import org.apache.cxf.ws.transfer.dialect.Dialect;
-import org.apache.cxf.ws.transfer.dialect.fragment.faults.InvalidExpression;
 import org.apache.cxf.ws.transfer.dialect.fragment.faults.UnsupportedLanguage;
+import org.apache.cxf.ws.transfer.dialect.fragment.language.FragmentDialectLanguage;
+import org.apache.cxf.ws.transfer.dialect.fragment.language.FragmentDialectLanguageQName;
+import org.apache.cxf.ws.transfer.dialect.fragment.language.FragmentDialectLanguageXPath10;
 import org.apache.cxf.ws.transfer.shared.TransferTools;
 import org.apache.cxf.ws.transfer.shared.faults.UnknownDialect;
 
@@ -49,27 +45,12 @@ import org.apache.cxf.ws.transfer.shared.faults.UnknownDialect;
  */
 public class FragmentDialect implements Dialect {
     
-    private static Pattern qNamePattern;
+    private final Map<String, FragmentDialectLanguage> languages;
     
     public FragmentDialect() {
-        if (qNamePattern == null) {
-            // See http://www.w3.org/TR/REC-xml-names/#ns-decl
-            // prefix name start char
-            String prefixNameStartChar = "[A-Z]|_|[a-z]|[\\x{c0}-\\x{d6}]|[\\x{d8}-\\x{f6}]|[\\x{f8}-\\x{2ff}]|"
-                                 + "[\\x{370}-\\x{37d}]|[\\x{37f}-\\x{1fff}]|[\\x{200c}-\\x{200d}]|"
-                                 + "[\\x{2070}-\\x{218f}]|[\\x{2c00}-\\x{2fef}]|[\\x{3001}-\\x{d7ff}]|"
-                                 + "[\\x{f900}-\\x{fdcf}]|[\\x{fdf0}-\\x{fffd}]|[\\x{10000}-\\x{effff}]";
-            // xml name start char
-            String nameStartChar = ":|" + prefixNameStartChar;
-            // prefix continue chars
-            String prefixNameChar = "-|\\.|[0-9]|\\x{b7}|[\\x{0300}-\\x{036f}]|[\\x{203f}-\\x{2040}]"
-                                  + "|" + prefixNameStartChar;
-            // xml name continue chars
-            String nameChar = ":|" + prefixNameChar;
-            String qName = String.format("^((%s)(%s)*:)?(%s)(%s)*",
-                                         prefixNameStartChar, prefixNameChar, nameStartChar, nameChar);
-            qNamePattern = Pattern.compile(qName);
-        }
+        languages = new HashMap<String, FragmentDialectLanguage>();
+        languages.put(FragmentDialectConstants.QNAME_LANGUAGE_IRI, new FragmentDialectLanguageQName());
+        languages.put(FragmentDialectConstants.XPATH10_LANGUAGE_IRI, new FragmentDialectLanguageXPath10());
     }
 
     @Override
@@ -77,10 +58,9 @@ public class FragmentDialect implements Dialect {
         for (Object o : body.getAny()) {
             if (o instanceof JAXBElement && ((JAXBElement)o).getValue() instanceof ExpressionType) {
                 ExpressionType expression = (ExpressionType) ((JAXBElement)o).getValue();
-                if (FragmentDialectConstants.QNAME_LANGUAGE_IRI.equals(expression.getLanguage())) {
-                    return generateGetResponse(processGetQName(representation, expression));
-                } else if (FragmentDialectConstants.XPATH10_LANGUAGE_IRI.equals(expression.getLanguage())) {
-                    return generateGetResponse(processGetXPath10(representation, expression));
+                if (languages.containsKey(expression.getLanguage())) {
+                    FragmentDialectLanguage language = languages.get(expression.getLanguage());
+                    return generateGetResponse(language.getResourceFragment(representation, expression));
                 } else {
                     throw new UnsupportedLanguage();
                 }
@@ -105,101 +85,27 @@ public class FragmentDialect implements Dialect {
         throw new UnknownDialect();
     }
     
-    private NodeList processGetQName(final Representation representation, ExpressionType expression) {
-        try {
-            String expressionStr = getQNameXPathFromExpression(expression);
-            // Evaluate XPath
-            XPath xPath = TransferTools.getXPath();
-            xPath.setNamespaceContext(new NamespaceContext() {
-
-                @Override
-                public String getNamespaceURI(String prefix) {
-                    if (prefix != null && !prefix.isEmpty()) {
-                        Element resource = (Element) representation.getAny();
-                        return resource.getAttribute("xmlns:" + prefix);
-                    } else {
-                        return null;
-                    }
-                }
-
-                @Override
-                public String getPrefix(String string) {
-                    throw new UnsupportedOperationException();
-                }
-
-                @Override
-                public Iterator getPrefixes(String string) {
-                    throw new UnsupportedOperationException();
-                }
-            });
-            return (NodeList) xPath.evaluate(
-                    expressionStr, representation.getAny(), XPathConstants.NODESET);
-        } catch (XPathExpressionException ex) {
-            throw new RuntimeException(ex);
+    /**
+     * Register FragmentDialectLanguage object for IRI.
+     * @param iri
+     * @param language 
+     */
+    public void registerLanguage(String iri, FragmentDialectLanguage language) {
+        if (languages.containsKey(iri)) {
+            throw new IllegalArgumentException(String.format("IRI \"%s\" is already registered", iri));
         }
+        languages.put(iri, language);
     }
     
-    private String getQNameXPathFromExpression(ExpressionType expression) {
-        if (expression.getContent().size() == 1) {
-            String expressionValue = (String) expression.getContent().get(0);
-            Matcher m = qNamePattern.matcher(expressionValue);
-            if (m.matches()) {
-                return "/node()/" + expressionValue;
-            } else {
-                throw new InvalidExpression();
-            }
-        } else {
-            throw new InvalidExpression();
+    /**
+     * Unregister FragmentDialectLanguage object for IRI.
+     * @param iri 
+     */
+    public void unregisterLanguage(String iri) {
+        if (!languages.containsKey(iri)) {
+            throw new IllegalArgumentException(String.format("IRI \"%s\" is not registered", iri));
         }
-    }
-    
-    private Object processGetXPath10(final Representation representation, ExpressionType expression) {
-        String expressionStr = getXPathFromExpression(expression);
-        // Evaluate XPath
-        XPath xPath = TransferTools.getXPath();
-        xPath.setNamespaceContext(new NamespaceContext() {
-
-            @Override
-            public String getNamespaceURI(String prefix) {
-                if (prefix != null && !prefix.isEmpty()) {
-                    Element resource = (Element) representation.getAny();
-                    return resource.getAttribute("xmlns:" + prefix);
-                } else {
-                    return null;
-                }
-            }
-
-            @Override
-            public String getPrefix(String string) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public Iterator getPrefixes(String string) {
-                throw new UnsupportedOperationException();
-            }
-        });
-        try {
-            return (Node) xPath.evaluate(
-                expressionStr, representation.getAny(), XPathConstants.NODE);
-        } catch (XPathException ex) {
-            // See https://www.java.net/node/681793
-        }
-
-        try {
-            return (String) xPath.evaluate(
-                expressionStr, representation.getAny(), XPathConstants.STRING);
-        } catch (XPathException ex) {
-            throw new InvalidExpression();
-        }
-    }
-    
-    private String getXPathFromExpression(ExpressionType expression) {
-        if (expression.getContent().size() == 1) {
-            return (String) expression.getContent().get(0);
-        } else {
-            throw new InvalidExpression();
-        }
+        languages.remove(iri);
     }
     
     private JAXBElement<ValueType> generateGetResponse(Object value) {
